Tugas coding web PHP minggu - 12
coding php dan bootstrap

+++ restoran/kategori/delete.php
@@ -1,13 +1,16 @@
 <?php 
 
-    require_once "../function.php";
+if (isset($_GET['id'])) {
+    $id=$_GET['id'];
+
+    $sql = "DELETE FROM tblkategori WHERE idkategori=$id";
+
+    $db->runSQL($sql);
+
+    header("location:?f=kategori&m=select");
 
     
+}
 
-    $sql = "DELETE FROM tblkategori WHERE idkategori = $id";
-
-     $hasil = mysqli_query($koneksi, $sql);
-
-    header("location:http://localhost/phpsmk-tugasvideo-1-55/phpsmk-tugasvideo-1-55/restoran/kategori/select.php?");
 
 ?>
